Fix inheritance.
The way we receive PTRACE_EVENT_FORK from parent and SIGSTOP from child
looks totally random. This looks like a ptrace bug but i think this is
the way to solve it. Here's what we do:

Keep if the child has inherited sandbox data in child->inherited.

When we receive a SIGSTOP from a child *before* we receive
PTRACE_EVENT_FORK from her parent, add the child to the list of
children, set it up but do *not* resume it until we receive the
PTRACE_EVENT_FORK.

When we receive PTRACE_EVENT_FORK, get the new child's process id using
ptrace(PTRACE_GETEVENTMSG,...) and check if she has already been added
to the list of children. If the child hasn't been added to the list of
children, add it, do the inheritance, set it up and resume it. If the
child is already in the list of children but hasn't yet inherited
sandbox data, inherit parent's sandbox data and resume it.