Implement -W/--wait-all
This command line flag makes sydbox wait for all children, not just the
eldest, to exit before exiting. This doesn't work for exhereses because
they may execute commands in pkg_postinst like udevd --daemon and sydbox
will cause the installation process to hang because it'll wait for that
process to exit.